vehicles reorg: move display code into own file
Move all the display related stuff into a separate file for easier
maintenance.
